Login 관련된 폴덩정리 및 post 와 Member CRUD 결합

+++ src/main/java/com/example/waffleeungaebackend/service/PostServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.waffleeungaebackend.service;
 
+import com.example.waffleeungaebackend.dto.MemberDto;
 import com.example.waffleeungaebackend.dto.PostDto;
-import com.example.waffleeungaebackend.dto.SessionMember;
 import com.example.waffleeungaebackend.dto.request.PostCreateRequestDto;
 import com.example.waffleeungaebackend.dto.request.PostPatchRequestDto;
 import com.example.waffleeungaebackend.entity.Category;
@@ -17,6 +17,7 @@ import org.springframework.stereotype.Service;
 
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 
@@ -38,17 +39,22 @@ public class PostServiceImpl implements PostService{
 
 
     @Override
-    public Post addPostList(PostCreateRequestDto postCreateRequestDto, Long categoryId, String  memberName) {
-        Member member = memberRepository.findByName(memberName);
+    public Post addPostList(PostCreateRequestDto postCreateRequestDto, Long categoryId, Long  memberId) {
+        Member member = memberRepository.findById(memberId).orElseThrow(() -> new IllegalArgumentException("유저정보가 없습니다."));
         postCreateRequestDto.setMember(member);
         Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new IllegalArgumentException("TODO 생성실패"));
         return postRepository.save(postCreateRequestDto.toEntity(category));
     }
 
     @Override
-    public Post updateById(Long id, PostPatchRequestDto postPatchRequestDto){
+    public Post updateById(Long id, PostPatchRequestDto postPatchRequestDto, Long memberID){
+        Member member = memberRepository.findById(memberID).orElseThrow(() -> new IllegalArgumentException("유저정보가 없습니다."));
         Post post = this.findById(id);
         Category category = categoryRepository.findById(postPatchRequestDto.getCategoryId()).orElseThrow(() -> new IllegalArgumentException("TODO 생성실패"));
+
+        if (!Objects.equals(post.getMember().getMemberId(), memberID)) {
+            throw new IllegalStateException("해당 글을 삭제할 권한이 없습니다.");
+        }
 
         post.setCategory(category);
         post.setContent(postPatchRequestDto.getContent());
@@ -57,7 +63,12 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public void deletePostList(Long id) {
+    public void deletePostList(Long id, Long memberID) {
+        Member member = memberRepository.findById(memberID).orElseThrow(() -> new IllegalArgumentException("유저정보가 없습니다."));
+        Post post = this.findById(id);
+        if (!Objects.equals(post.getMember().getMemberId(), memberID)) {
+            throw new IllegalStateException("해당 글을 삭제할 권한이 없습니다.");
+        }
         postRepository.deleteById(id);
     }
